+++ trellis/containers/BUILD
@@ -1,17 +1,15 @@
 cc_library(
     name = "containers",
+    srcs = [
+        "transforms.cpp",
+    ],
     hdrs = [
         "fifo.hpp",
         "multi_fifo.hpp",
+        "transforms.hpp",
     ],
     visibility = ["//visibility:public"],
-)
-
-cc_test(
-    name = "test",
-    srcs = ["test/test_multi_fifo.cpp"],
     deps = [
-        ":containers",
-        "@gtest//:gtest_main",
+        "//trellis/core:core_time",
     ],
 )
