+++ trellis/containers/test/BUILD
@@ -1,0 +1,20 @@
+cc_test(
+    name = "test_container_transforms",
+    srcs = [
+        "test_transforms.cpp",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//trellis/containers",
+        "@gtest//:gtest_main",
+    ],
+)
+
+cc_test(
+    name = "test_multi_fifo",
+    srcs = ["test_multi_fifo.cpp"],
+    deps = [
+        "//trellis/containers",
+        "@gtest//:gtest_main",
+    ],
+)
